% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{call_api_once}
\alias{call_api_once}
\title{Call an API once}
\usage{
call_api_once(URL, path, query, ...)
}
\arguments{
\item{URL}{endpoint url}

\item{path}{the path}

\item{query}{the actual query (with list())}

\item{...}{whichever to add behind path}
}
\description{
Call an API once
}
