% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_data}
\alias{get_data}
\title{Get data from the Flemish parliament}
\usage{
get_data(
  type = "speech",
  fact = "debates",
  date_range_from,
  date_range_to,
  plen_comm = "plen",
  use_parallel = FALSE,
  raw = FALSE
)
}
\arguments{
\item{type}{Type of data to be returned, options include "document", "speech" or "basicdata".}

\item{fact}{Which fact should be returned, options include "written_questions", "debates", "oral_questions_and_interpellations", "parliamentary_initiatives" or "council_hearings"}

\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{plen_comm}{Switch to pick between plenary (plen) and commission (comm) sessions.}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{raw}{Boolean: should the raw object be returned?}
}
\description{
Get data from the Flemish parliament
}
\examples{

\dontrun{

 get_data(date_range_from="2022-01-20",
             date_range_to="2022-01-31",
             type="document",
             fact="written_questions",
             use_parallel=TRUE )

}
}
