% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_legislatures}
\alias{get_legislatures}
\title{Get current and past legislatures of the Flemish parliament}
\usage{
get_legislatures()
}
\description{
Get current and past legislatures of the Flemish parliament
}
\examples{
\dontrun{
get_legislatures()
}
}
