% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_mp}
\alias{get_mp}
\title{Search all the MPs serving, or who served the Flemish parliament}
\usage{
get_mp(
  selection = "current",
  fact = "bio",
  date_at = NULL,
  use_parallel = TRUE
)
}
\arguments{
\item{selection}{Select either "current", at a certain date "date" or "former".}

\item{fact}{Options of facts to return are 'raw', 'bio', 'education', 'career','political_info', 'presences_commissions' or 'presences_plenary'.}

\item{date_at}{When selecting "date" in selection, provide the date, using the following format "yyyy-mm-dd".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}
}
\description{
Search all the MPs serving, or who served the Flemish parliament
}
\examples{
\dontrun{

 get_mp(use_parallel=TRUE,fact="bio",date_at="1998-01-01",selection="date")

}
}
