% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_plen_comm_details}
\alias{get_plen_comm_details}
\title{Get the details from plenary or commission sessions}
\usage{
get_plen_comm_details(
  date_range_from,
  date_range_to,
  fact,
  plen_comm = "plen",
  type = "details",
  use_parallel = TRUE,
  raw = FALSE,
  extra_via_fact = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{fact}{The fact to search}

\item{plen_comm}{Switch to pick between plenary (plen) and commission (comm) sessions.}

\item{type}{Type of data to be returned, options include "document", "speech" or "details".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{raw}{Boolean: should the raw object be returned?}

\item{extra_via_fact}{Boolean: also search the underlying endpoint for linked documents? This may return documents not linked to the specific meeting, thus may also include meetings on dates before/after the date range.}
}
\description{
Get the details from plenary or commission sessions
}
