% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_plen_comm_documents}
\alias{get_plen_comm_documents}
\title{Get the documents discussed in plenary or commission sessions}
\usage{
get_plen_comm_documents(
  date_range_from,
  date_range_to,
  fact,
  plen_comm = "plen",
  use_parallel = TRUE,
  raw = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{fact}{The fact to search.}

\item{plen_comm}{Switch to pick between plenary (plen) and commission (comm) sessions.}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{raw}{Boolean: should the raw object be returned?}
}
\description{
Get the documents discussed in plenary or commission sessions
}
