% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_sessions_details}
\alias{get_sessions_details}
\title{Get the details of sessions}
\usage{
get_sessions_details(
  date_range_from,
  date_range_to,
  plen_comm = "plen",
  use_parallel = TRUE,
  raw = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{plen_comm}{Switch to pick between plenary (plen) and commission (comm) sessions.}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{raw}{Boolean: should the raw object be returned?}
}
\description{
Get the details of sessions
}
\examples{
\dontrun{

get_sessions_details(date_range_from="2021-09-01"
                     , date_range_to="2021-12-31"
                     , plen_comm = "plen"
                     , use_parallel=TRUE)

}
}
