% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_work}
\alias{get_work}
\title{Get data from the Flemish parliament}
\usage{
get_work(
  date_range_from,
  date_range_to,
  fact = "debates",
  type = "details",
  plen_comm = "plen",
  use_parallel = TRUE,
  raw = FALSE,
  extra_via_fact = FALSE,
  two_columns_pdf = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{fact}{Which fact should be returned, options include "written_questions", "debates", "oral_questions_and_interpellations", "parliamentary_initiatives" or "committee_hearings"}

\item{type}{Type of data to be returned, options include "document", "speech" or "details".}

\item{plen_comm}{Switch to pick between plenary (plen) and commission (comm) sessions.}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{raw}{Boolean: should the raw object be returned?}

\item{extra_via_fact}{Boolean: also search the underlying endpoint for linked documents? This may return documents not linked to the specific meeting, thus may also include meetings on dates before/after the date range.}

\item{two_columns_pdf}{Boolean: Use when you encounter two-colummned PDFs.}
}
\description{
Get data from the Flemish parliament
}
\examples{
\dontrun{

get_work(
  date_range_from = "2022-01-20",
  date_range_to = "2022-01-31",
  type = "document",
  fact = "written_questions",
  use_parallel = TRUE
)
}
}
