% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_written_questions_basicdata}
\alias{get_written_questions_basicdata}
\title{Get the basic data of the written questions}
\usage{
get_written_questions_basicdata(
  date_range_from,
  date_range_to,
  use_parallel = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}
}
\description{
Get the basic data of the written questions
}
