% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_written_questions_details}
\alias{get_written_questions_details}
\title{Get the details of the written questions}
\usage{
get_written_questions_details(
  date_range_from,
  date_range_to,
  use_parallel = TRUE,
  raw = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{raw}{Boolean: should the raw object be returned?}
}
\description{
Get the details of the written questions
}
