% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{guarantee_field}
\alias{guarantee_field}
\title{Guarantee that certain variables are present}
\usage{
guarantee_field(df, fields)
}
\arguments{
\item{df}{dataframe}

\item{fields}{the field you want to guarantee}
}
\description{
Guarantee that certain variables are present
}
