% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{parse_documents}
\alias{parse_documents}
\title{Parse PDF documents}
\usage{
parse_documents(mainlist, use_parallel = TRUE, two_columns_pdf = FALSE)
}
\arguments{
\item{mainlist}{Data frame with documents to be parsed}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{two_columns_pdf}{Boolean: Use when you encounter two-colummned PDFs}
}
\description{
Parse PDF documents
}
