% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{read_text}
\alias{read_text}
\title{Function to read in two-column pdfs}
\usage{
read_text(text)
}
\arguments{
\item{text}{result from pdftools call}
}
\description{
Function to read in two-column pdfs
}
