% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{remove_trailing_and_leading_slashes}
\alias{remove_trailing_and_leading_slashes}
\title{Remove trailing and preceding / from a given string}
\usage{
remove_trailing_and_leading_slashes(input_string)
}
\arguments{
\item{input_string}{input string}
}
\description{
Remove trailing and preceding / from a given string
}
