% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{search_mp}
\alias{search_mp}
\title{Search all the MPs serving, or who served the Flemish parliament}
\usage{
search_mp(
  selection = "current",
  fact = "bio",
  date_at = NULL,
  use_parallel = FALSE
)
}
\arguments{
\item{selection}{select either "current", at a certain date ("date_at") or "former"}

\item{fact}{Options of facts to return are 'raw', 'bio', 'education', 'career','political_info', 'presences_commissies' or 'presences_plenary'}

\item{date_at}{When selecting "date_at" in selection, provide date, should be in format "yyyy-mm-dd".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}
}
\description{
Search all the MPs serving, or who served the Flemish parliament
}
\examples{
\dontrun{

 search_mp(use_parallel=TRUE,fact="bio",date_at="1998-01-01",selection="date")

}
}
