% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_plenary.R
\name{search_plenary}
\alias{search_plenary}
\title{Search for plenary sessions}
\usage{
search_plenary(date_range_from, date_range_to, use_parallel = FALSE)
}
\arguments{
\item{date_range_from}{The startdate, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The enddate, should be in format "yyyy-mm-dd".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}
}
\value{
Returns a dataframe with IDs of the plenary sessions in the indicated date range, together with the journal IDs and activtity IDs.
}
\description{
Search for plenary sessions of the Flemish Parliament between certain dates.
}
\examples{

\dontrun{
plenary_object <-  vlaplr::search_plenary(date_range_from = "2021-01-01",
                                      date_range_to= "2021-03-31",
                                      use_parallel=TRUE)
View(plenary_object)
}
}
