% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{search_terms}
\alias{search_terms}
\title{Filter texts fields on certain search terms}
\usage{
search_terms(df, text_field, search_terms = NULL)
}
\arguments{
\item{df}{Data frame}

\item{text_field}{Which textfield should be searched?}

\item{search_terms}{The search terms. Multiple terms are possible by adding them as a vector with c(). This is not case sensitive.}
}
\description{
Filter texts fields on certain search terms
}
\examples{
\dontrun{

wq_document <- get_work(
  date_range_from = "2022-01-20",
  date_range_to = "2022-01-31",
  type = "document",
  fact = "written_questions",
  use_parallel = TRUE
)

search_terms(wq_document, text_field = "text", search_terms = "PFOS")
}
}
