% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_sessions_details}
\alias{get_sessions_details}
\title{Get the details of sessions}
\usage{
get_sessions_details(
  date_range_from,
  date_range_to,
  plen_comm = "plen",
  use_parallel = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{plen_comm}{Switch to pick between plenary (plen) and commission (comm) sessions.}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}
}
\description{
Get the details of sessions
}
