% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{get_written_questions_documents}
\alias{get_written_questions_documents}
\title{Get the PDF documents of the written questions}
\usage{
get_written_questions_documents(
  date_range_from,
  date_range_to,
  use_parallel = TRUE,
  two_columns_pdf = FALSE
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}

\item{two_columns_pdf}{Boolean: Use when you encounter two-colummned PDFs}
}
\description{
Get the PDF documents of the written questions
}
