% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_terms.R
\name{search_terms}
\alias{search_terms}
\title{Search for specific terms}
\usage{
search_terms(plenary_object, search_terms, type = "all", use_parallel = FALSE)
}
\arguments{
\item{plenary_object}{A plenary object created by search_plenary().}

\item{search_terms}{The search terms. Multiple terms are possible by adding them as a vector with c(). This is not case sensitive.}

\item{type}{Which type of activity do you want to search? Valid options are "all", "debatten","gedachtenwisselingen","vragen_interpelaties" and "parlementaire_initiatieven".}

\item{use_parallel}{Boolean: should parallel workers be used to call the API?}
}
\value{
Dataframe with the full text and the author.
}
\description{
Search for specific terms in all spoken word during the plenary sessions in the Flemish parliament.
Requires a plenary_object created with vlaplr::search_plenary().
}
\examples{
\dontrun{
plenary_object <- search_plenary(date_range_from = "2021-01-01",
                                      date_range_to= "2021-03-31",
                                      use_parallel=TRUE)
search_terms(plenary_object = plenary_object,
             search_terms =  c("statistiek","welzijn"),
             type = c("all"),
             use_parallel=TRUE )
}
}
