% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_script.R
\name{use_generalized_query}
\alias{use_generalized_query}
\title{Get data by using the native query engine}
\usage{
use_generalized_query(
  date_range_from,
  date_range_to,
  type = "Schriftelijke vraag"
)
}
\arguments{
\item{date_range_from}{The start date, should be in format "yyyy-mm-dd".}

\item{date_range_to}{The end date, should be in format "yyyy-mm-dd".}

\item{type}{Vector of types, see get_all_agg_types() for all options}
}
\description{
Get data by using the native query engine
}
